extract findBulkById
